# Conflict resolved:
#	collect_app/src/main/java/org/odk/collect/android/widgets/AnnotateWidget.java
#	collect_app/src/main/java/org/odk/collect/android/widgets/ArbitraryFileWidget.java
#	collect_app/src/main/java/org/odk/collect/android/widgets/ImageWidget.java
#	gradle/wrapper/gradle-wrapper.properties

+++ collect_app/src/main/java/org/odk/collect/android/widgets/VideoWidget.java
@@ -28,7 +28,6 @@ import android.preference.PreferenceManager;
 import android.provider.MediaStore.Video;
 import android.support.annotation.NonNull;
 import android.view.View;
-import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.Toast;
@@ -197,7 +196,7 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         MediaManager
                 .INSTANCE
                 .markOriginalFileOrDelete(getFormEntryPrompt().getIndex().toString(),
-                getInstanceFolder() + File.separator + binaryName);
+                        getInstanceFolder() + File.separator + binaryName);
         binaryName = null;
     }
 
@@ -219,23 +218,40 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         }
     }
 
+    /**
+     * Set this widget with the actual file returned by OnActivityResult.
+     * Both of Uri and File are supported.
+     * If the file is local, a Uri is enough for the copy task below.
+     * If the chose file is from cloud(such as Google Drive),
+     * The retrieve and copy task is already executed in the previous step,
+     * so a File object would be presented.
+     *
+     * @param object Uri or File of the chosen file.
+     * @see org.odk.collect.android.activities.FormEntryActivity#onActivityResult(int, int, Intent)
+     */
     @Override
-    public void setBinaryData(Object binaryuri) {
-        if (binaryuri == null || !(binaryuri instanceof Uri)) {
-            Timber.w("AudioWidget's setBinaryData must receive a Uri object.");
+    public void setBinaryData(Object object) {
+        File newVideo = null;
+        // get the file path and create a copy in the instance folder
+        if (object instanceof Uri) {
+            String sourcePath = getSourcePathFromUri((Uri) object);
+            String destinationPath = getDestinationPathFromSourcePath(sourcePath);
+            File source = fileUtil.getFileAtPath(sourcePath);
+            newVideo = fileUtil.getFileAtPath(destinationPath);
+            fileUtil.copyFile(source, newVideo);
+        } else if (object instanceof File) {
+            newVideo = (File) object;
+        } else {
+            Timber.w("VideoWidget's setBinaryData must receive a File or Uri object.");
             return;
         }
 
-        // get the file path and create a copy in the instance folder
-        Uri uri = (Uri) binaryuri;
 
-        String sourcePath = getSourcePathFromUri(uri);
-        String destinationPath = getDestinationPathFromSourcePath(sourcePath);
+        if (newVideo == null) {
+            Timber.e("setBinaryData FAILED");
+            return;
+        }
 
-        File source = fileUtil.getFileAtPath(sourcePath);
-        File newVideo = fileUtil.getFileAtPath(destinationPath);
-
-        fileUtil.copyFile(source, newVideo);
 
         if (newVideo.exists()) {
             // Add the copy to the content provier
@@ -270,10 +286,11 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         // a bug in the Nexus 7 on 4.3 not returning the mediaUri in the data
         // of the intent - uri in this case is a file
         if (NEXUS7.equals(MODEL) && Build.VERSION.SDK_INT == 18) {
-            File fileToDelete = new File(uri.getPath());
-            int delCount = fileToDelete.delete() ? 1 : 0;
-
-            Timber.i("Deleting original capture of file: %s count: %d", uri.toString(), delCount);
+            if (object instanceof File) {
+                File fileToDelete = (File) object;
+                int delCount = fileToDelete.delete() ? 1 : 0;
+                Timber.i("Deleting original capture of file: %s count: %d", binaryName, delCount);
+            }
         }
     }
 
@@ -295,14 +312,6 @@ public class VideoWidget extends QuestionWidget implements FileWidget {
         String extension = sourcePath.substring(sourcePath.lastIndexOf('.'));
         return getInstanceFolder() + File.separator
                 + fileUtil.getRandomFilename() + extension;
-    }
-
-    @Override
-    public void setFocus(Context context) {
-        // Hide the soft keyboard if it's showing.
-        InputMethodManager inputManager = (InputMethodManager) context
-                .getSystemService(Context.INPUT_METHOD_SERVICE);
-        inputManager.hideSoftInputFromWindow(this.getWindowToken(), 0);
     }
 
     @Override
